resource: Refactor jobinfo data structure
